mcp_server/requirements.txt
annotated-types==0.7.0
cachetools==5.5.2
certifi==2025.1.31
click==8.1.8
google-auth==2.38.0
google-genai==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
mcp==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.11.0b1
pydantic-settings==2.8.1
pydantic_core==2.31.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9
setuptools==76.0.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.6.6
uvicorn==0.34.0
websockets==14.2
wheel==0.45.1
asyncpg==0.27.0
psycopg[pool]==3.1.8
# Cloud SQL dependencies for JSON credential authentication
google-auth==2.38.0
google-cloud-sql-connector==1.0.0
pg8000==1.30.3